+++ src/benchmark/frontend/src/views/Config.vue
@@ -901,9 +901,10 @@ function normalizeBuildParams(configs) {
           :items="[
             { key: 'GGML_CUDA_ENABLE_UNIFIED_MEMORY', label: 'GGML_CUDA_ENABLE_UNIFIED_MEMORY', type: 'text' },
             { key: 'CUDA_SCALE_LAUNCH_QUEUES', label: 'CUDA_SCALE_LAUNCH_QUEUES', type: 'select' },
-            { key: 'GGML_CUDA_P2P', label: 'GGML_CUDA_P2P', type: 'text' },
+            { key: 'GGML_CUDA_P2P', label: 'GGML_CUDA_P2P', type: 'select' },
           ]"
           :queue-options="['1x', '4x', '8x']"
+          :p2p-options="['off', 'on']"
           v-model="visualConfigs.export_configs"
         />